modification to entry.s
Here we are checking the value of the mhartid
register. This should return zero given our minimal
setup and knowledge from the spec in regard that at
least one hart must have an integer ID of zero. If
the read is not equal to zero we loopforever, other-
wise we print the ascii character 'U'.